+++ inc/woocommerce/integrations.php
@@ -97,6 +97,14 @@ function storefront_woocommerce_integrations_scripts() {
 		wp_enqueue_style( 'storefront-woocommerce-bundles-style', get_template_directory_uri() . '/inc/woocommerce/css/bundles.css', 'storefront-woocommerce-style' );
 		wp_style_add_data( 'storefront-woocommerce-bundles-style', 'rtl', 'replace' );
 	}
+
+	/**
+	 * WooCommerce Multiple Shipping Addresses
+	 */
+	if ( is_woocommerce_extension_activated( 'WC_Ship_Multiple' ) ) {
+		wp_enqueue_style( 'storefront-woocommerce-sma-style', get_template_directory_uri() . '/inc/woocommerce/css/ship-multiple-addresses.css', 'storefront-woocommerce-style' );
+		wp_style_add_data( 'storefront-woocommerce-sma-style', 'rtl', 'replace' );
+	}
 }
 
 /**